[vllm] feat: enable FSDP and vLLM 0.8.2 to support DPSK v3 training. (#923)
This is a solution to an error that occurs when vLLM 0.8.2 loads models
in the ds_v3 MoE format. For MoE models, directly calling vLLM's
`load_weights· function will result in an error, primarily because the
model params lose the related methods of the `FusedMoe` class.
Therefore, it is necessary to identify and call the specific parameter
loading method based on the model parameter names at runtime. However,
the current version is a dirty implementation because, in reality, if
invasive changes were made to vLLM, it would only require modifying 2
lines of code and adding a new function to identify the layer count (I
have already marked comments in the code). I’m not sure if there’s a
better implementation and would like some suggestions.