[refactor][single_controller] Small refactor and fixes in `worker.py` and `ray.base.py`  (#1470)
### Checklist Before Starting

- [x] Search for similar PR(s).

### What does this PR do?

1. The existing `WorkerHelper::_get_pid()` returns nothing; this PR fix
this nit by returning `os.getpid()`;
2. The `WorkerMeta` class in `worker.py` is only used in `Worker.init()`
and no where else. This class just maintains a list of env keys and
wraps a dict named `store`. This PR delete this class and move the
contents inside the `Worker` class. In this way, it would be easier if a
user want to subclass Worker, with different env keys;
3. In `merge_resource_pool` function, instead of directly return a
`RayResourcePool`, this PR changes the return type to be `type(rp1)`. In
this way, the function can be applied to not only `RayResourcePool`, but
also any subclass of `RayResourcePool`.

### Note

This PR splits off some small nits and refactors from #1454, so that the
small things here could be reviewed and merged sooner before we decide
on the structural PR #1454


### Checklist Before Submitting

- [x] Read the [Contribute
Guide](https://github.com/volcengine/verl?tab=readme-ov-file#contribution-guide).
- [x] Apply [pre-commit
checks](https://github.com/volcengine/verl?tab=readme-ov-file#code-linting-and-formatting).
- [ ] Add `[BREAKING]` to the PR title if it breaks any API.
- [ ] Update the documentation about your changes in the
[docs](https://github.com/volcengine/verl/tree/main/docs).
- [ ] Add CI test(s) if neccessary.

---------

Signed-off-by: Hongpeng Guo <hg5@illinois.edu>