[megatron] feat: Config NCCL Timeout for Megatron Backend Model Loading (#1983)
### Checklist Before Starting

- [x] Searched for similar PR(s).

### What does this PR do?

> This merge request addresses an issue encountered when using Megatron
as the backend for loading models with
`load_state_dict_to_megatron_gptmodel`. Specifically, when loading 32B
or larger models on 64 or more GPUs, it is common to exceed the default
NCCL timeout of 10 minutes(default 10 mins for
[torch.distributed.init_process_group("nccl")](https://docs.pytorch.org/docs/stable/distributed.html),
leading to errors during the[ dist.barrier()
](https://github.com/none0663/verl/blob/12f98137d54ef1e7b727022a10733d8fcc14986d/verl/models/mcore/loader.py#L463)call.


https://github.com/none0663/verl/blob/12f98137d54ef1e7b727022a10733d8fcc14986d/verl/models/mcore/loader.py#L360


https://github.com/none0663/verl/blob/12f98137d54ef1e7b727022a10733d8fcc14986d/verl/models/mcore/loader.py#L463


To mitigate this issue, this PR introduces a configuration option to
increase the NCCL timeout. This enhancement allows users to easily
adjust the timeout duration when encountering errors, improving the
robustness of model loading in distributed settings.

Thank you for considering this change!