[recipe] chore: entropy removes `WANDB_API_KEY` in code (#3956)
### What does this PR do?

It's a good idea to remove the setting of `WANDB_API_KEY` with some
placeholders like `YOUR_WANDB_API_KEY` in the code. Setting it is
something users usually keep in mind when they use wandb, and having
them in example bash scripts are already enough. It will be very hard
for users to notice, debug, and find out that there's another
placeholder in the code.